Deprecated base classes
Added deprecation annotations to TrackerActivityBase and
TrackerListActivityBase. Going forward, use of these base classes is
discouraged in favour of the simplified initialization API
(TrackingContext.Instance().start()).

This commit also moves the custom logo in the custom invite sample to the
root of the src directory so that it can be picked up.

Also fixes the packaging script (was copying jars into sample project dirs
that are no longer in use).

Fix for invalid SID in custom-invite sample.

Sample apps configured with a repeatDaysAfterDecline value.